[#11290] logback-it, extract TestBase
